[MDC-1500][MDC-1504] deb: random delay for cron.daily, cron.weekly
cron.daily/weekly anacron-style jobs have their start time not
randomized, and the various jobs are run sequentially.  Refer to
run-parts(8) for details.

That means we cannot simply sleep for a (long) random time inside such a
job, as it is also delaying subsequent jobs.

Unfortunately, simply detaching part of the job to the background is not
the solution, either: the background job will continue running in the
same process session as run-parts, and might be terminated (even if it
sets SIGHUP to IGNORE) when run-parts finishes execution.

On the other hand, we simply *cannot* tolerate the group sync behavior
we get from cron.daily and cron.weekly on tens of thousands of devices
attempting to connect at the same time.

As a compromise, assume a random delay of up to five minutes is
tolerable for cron.daily and cron.weekly job sequencing.  If that proves
to be too long, even two minutes or one minute would already be a major
help.

* Add a random delay of up to 5 minutes (300 seconds) to cron.daily

* Change the random delay of cron.weekly to match the implementation
  in cron.daily, and also set it to 5 minutes.