[MDC-1258] simet-api: simet_report-queue-flush: fix curl, send compressed
The (unsent) LMAP report queue has its contents stored gzip-compressed,
and the SIMET LMAP collector has accepted gzip-encoded reports for a
while, now.

Switch simet_report-queue-flush.sh to send the already-compressed data
directly (with an appropriate Content-Encoding header).  This is faster,
uses less memory and CPU, and it also works around curl --retry issues
with the use of stdin for input data when a retry is required.