[MDC-1474] sspoofer: fix UDP checksum on non-ILP64
The code was incorrectly using the size of a pointer instead of the size
of the udp header struct when calculating the UDP header checksum.

This worked on ILP64 platforms by sheer luck (which threw off testing),
but it would just checksum half the header on ILP32 platforms.

Found by Coverity Scan.  Fixes CID 464731