twamp: avoid possibility of *NULL in twamp_report_statistics
The code is not supposed to *ever* call twamp_report_statistics() with a
NULL report structure, and twamp_report_statistics() has nothing to do
if report->result is NULL, which is also never supposed to happen in the
first place.

Check report and report->result for NULL at function start, and return
EINVAL in that case.

This is likely not fixing any active bug, but just because the code
would not call twamp_report_statistics() with a NULL report or NULL
report->result in the first place.

Fixes: CID 464409, CID 464360