tcpbw: remove unneded checks for NULL optarg
getopt() won't ever return an option character with a NULL optarg for
options declared with ":".  Checking for NULL optarg in that case is not
needed (and in fact we don't in most option handling code), and triggers
warnings from static analyzers about the discrepancy.

Remove the unneded (and harmless) checks for NULL optarg, as it is less
troublesome than adding uneeded (and harmless) checks for NULL optarg to
every option...