[MDC-1528] simet-api: actually refresh local schedules
Fix the detection of a locally-generated schedule.  Without this fix,
once generated, the local schedule would only be updated upon a
controller request, reboot, or package update.  With the fix, a new
local schedule is generated every time simet_lmap-fetch-schedule runs,
and loaded into lmapd if it differs from the previous schedule.

Fixes: commit b176958f5ec8e73aa6e0ad38c8341e7f27a47e37
       [MDC-1419] simet-api: multi-lmap instance controller support (partial)