[MDC-1533] inetup: disconnect from servers when MA is unregistered
The MA might lose its registration information for a while.  In that
case, we don't want inetupc to connect to the server clusters, but we
don't want to exit it either: it is far easier to reload it, than to
restart it.

Add a new WAIT_DISCONNECTED state to inetupc, where it does nothing but
wait for signals.  This state is normally reached through the DISCONNECT ->
DISCONNECT_WAIT state path, so it will idle wating for signals while
disconnected.  On a startup without credentials, it could be reached
directly from SIMET_INETUP_P_C_INIT / SIMET_INETUP_P_C_RECONNECT.

Add struct simet_inetup_server::is_shutdown to track what should be the
final state of transitions that go through DISCONNECT: WAIT_DISCONNECTED
(is_shutdown false) or SHUTDOWN (is_shutdown true).

Update uptimeserver_disconnect() and uptimeserver_disconnectwait() state
handlers to obey the is_shutdown flag, and add an is_shutdown parameter
to simet_uptime2_disconnect(), that sets it.

On startup and when reloading (SIGHUP), detect the lack of an agent-id
or an agent-token, and switch to the new WAIT_DISCONNECTED state.