[MDC-1534] inetup: enhance signal handling
Trigger a graceful exit on the first SIGINT, just like what is already
done for SIGQUIT and SIGTERM.  Use SA_RESETHAND for SIGINT, so that a
second SIGINT will abort the code if it is stuck.

Abort inetup if we fail to install the SIGHUP handler: it is needed to
trigger a reload of agent credentials, which we cannot do without.