[MDC-1533] inetup: agent-id and token files are mandatory
Previously, inetupc wouldn't start without an agent-id and agent-token.
Now, it does, but it won't do anything without them.

This means the CLI options that set the files where it should load the
agent-id and agent-token have been mandatory for quite a while, now.

Fix the CLI usage (help) text to that effect, and explicitly exit with a
BAD COMMANDLINE exit status if either option is missing, or empty.