[MDC-1437] simet-dnsaddr: prime parent domain measurement
Attempt to prime the parent domain data before carring out the
not-in-cache measurement.

The not-in-cache measurement is a lot slower when the parent domain data
is already cached in the resolver, and it is not possible to ensure this
situation at all.  Therefore, do the opposite: try to ensure that the
parent domain data is already cached in the resolver(s).

Since we might lose resolver IP information otherwise (a priming run
might train a multipath local resolver like systemd-resolved and thus
expose less servers to the non-cached/cached query-response latency
measurement), store the results of the priming as a new
"reflect-priming" subtype.

This change requires a SIMET2 data pipeline update, for the new
reflect-priming measurement subtype.  The updated data pipeline will
discard the timing measurements for "reflect-priming", but consume any
reflector IP addresses it detects.