[MDC-1437] simet-dnsaddr: improve error handling
Flag failures in sdnsa_get_reflect_domain() as an internal error, and
explicitly check for it.

For sdnsa_getaddrinfo(), check that node is neither NULL nor an empty
string (it was checking that node[9] is non-NUL, which is is useless
unless we also checked node[0..8] to be non-NUL).

Output an error message when sdnsa_render_report() fails, instead of
silently exiting with an error status.  Fix this error path to return
SEXIT_FAILURE instead of -errno as the program exit status, as well.