[MDC-1539] simet-dnsaddr: increase session-id length to 256 bits
Increase session-id length from 64 to 256 random bits.  This greatly
reduces the chances of a collision.

Note that since the random session-id goes through base64 *and* then
gets encoded as a DNS label which is case-insensitive, it won't have the
same colision resistente as a 256 bit random number, but it will still
be extremely strong (case insensitivity loses just one bit of colision
resistance per base64 symbol).