[MDC-1540] simet-dnsaddr: add dnssec measurement
Add the basic DNSSEC measurement: query a valid DNSSEC node in a signed
zone (must return a valid RR set), and query an invalid DNSSEC node in a
signed zone, must return an error.

It is slightly more complex than that because we do the queries a number
of times to measure query latency, and also because we try to prime the
resolver with the DNSSEC validation data for the parent zone.