[MDC-1540] simet-dnsaddr: add report column to LMAP results
Since a new more general result column will be needed to report DNSSEC
and system configuration data, the SIMET2 metric was revised to add such
a column (optional).

This new column holds a (serialized) JSON object.

Change sdnsa_render_reflect() to include the new column, using an empty
JSON object.  While the result column is (currently?) unused for the
reflect measurement subtypes, we will need it in place regardless, so
that the resulting table has no missing columns when a new measurement
subtype that does use it is added to it.

We could just leave the contents of the new column empty when unused,
but that makes it a bit more difficult to detect incorrect reports in
the data pipeline.