[MDC-1540] simet-dnsaddr: tolerate timeouts for dnssec
Don't abort the dnssec measurement on timeout (EAI_AGAIN): it is prone
to occour on the first query to BAD_NODE for unknown reasons, and since
we would bail out and drop the dnssec measurement on the first error,
this makes it impossible to even get enough data about the issue.

Change sdnsa_getaddrinfo() and sdnsa_getaddrinfo_error() to return -1 on
EAI_EAGAIN instead of SEXIT_FAILURE (positive values).  It will not
record the measurement, the only change is the return value.

For sdnsa_dnssec_query(), EAI_EAGAIN on any of the measurement steps
will still preclude further measurement steps, but it will not end the
measurement loop.