[MDC-1554] simet-api: interrupt report queue flushing on errors
Interrupt a queue flush run at the first error returned by the set of
LMAP collectors (with a few exceptions).

Without this change, simet_report-queue-flush.sh could trigger large
bursts of failed report delivery attempts, which is not nice for the
LMAP collectors.  Such a burst would get worse as the queue of
undelieverable reports grew.

With this change, all collectors are tried and if none accept the
report, simet_report-queue-flush.sh will interrupt the queue flush and
exit.  Note that it will not stop trying to flush the queue on
server-side (5xx) errors.

The reports are _not_ deleted from the queue, unless the LMAP collector
returns status 400 (delete this report, it is malformed) or status 410
(delete the entire queue).

This reduces the impact on the LMAP collectors on conditions that are
not going to be cleared just by retrying, and are not report-specific,
and status 410 gives the LMAP collector some control over out-of-order
MAs.

Note that status 410 is only honored if no other LMAP collector accepts
the report or returns status 400, 401 or 403.